pyproject.toml
[project]
name = "nexus"
version = "0.1.0"
description = "A configuration manager for python code."
authors = [
    {name = "Jorrit Vander Mynsbrugge",email = "44178217+jorritvm@users.noreply.github.com"}
]
license = {text = "Apache"}
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "pydantic (>=2.12.5,<3.0.0)",
    "update (>=0.0.1,<0.0.2)",
    "pyyaml (>=6.0.3,<7.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
